Allow references to omit the ".json" extension.
In order for JSON Schema references to work with RAMLCodeGen, the $ref value
must be the name of the schema used in the RAML file. However, SchemaToPoco
looks for a JSON file with the same name as the $ref value, which must include
the ".json" extension. This change allows the extension to be omitted, so that
the same schema file will work with both tools.

For best results, the schema name in the RAML file, the title in the schema
file, and the schema file name sans extension, should all match.